Updated to Zenilib v4.0.0 & Started adding in the GUI Code.  Currently Non-Functional however within a day or two a version should be uploaded which is functional

+++ trunk/src/Zeni/Collision.cpp
@@ -1,39 +1,25 @@
-/* This file is part of the Zenipex Library.
-* Copyleft (C) 2011 Mitchell Keith Bloch a.k.a. bazald
-*
-* The Zenipex Library is free software; you can redistribute it and/or 
-* modify it under the terms of the GNU General Public License as 
-* published by the Free Software Foundation; either version 2 of the 
-* License, or (at your option) any later version.
-*
-* The Zenipex Library is distributed in the hope that it will be useful, 
-* but WITHOUT ANY WARRANTY; without even the implied warranty of 
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU 
-* General Public License for more details.
-*
-* You should have received a copy of the GNU General Public License 
-* along with the Zenipex Library; if not, write to the Free Software 
-* Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 
-* 02110-1301 USA.
-*
-* As a special exception, you may use this file as part of a free software
-* library without restriction.  Specifically, if other files instantiate
-* templates or use macros or inline functions from this file, or you compile
-* this file and link it with other files to produce an executable, this
-* file does not by itself cause the resulting executable to be covered by
-* the GNU General Public License.  This exception does not however
-* invalidate any other reasons why the executable file might be covered by
-* the GNU General Public License.
-*/
+/* This file is part of the Zenipex Library (zenilib).
+ * Copyright (C) 2011 Mitchell Keith Bloch (bazald).
+ *
+ * zenilib is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * zenilib is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with zenilib.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
-#include <Zeni/Collision.hxx>
-
-#include <Zeni/Coordinate.hxx>
-#include <Zeni/Vector3f.hxx>
+#include <zeni.h>
 
 #include <cmath>
 
-#include <Zeni/Global.h>
+#include <Zeni/Define.h>
 
 namespace Zeni {
 
@@ -54,7 +40,7 @@ namespace Zeni {
     }
 
     inline void absolute_float_unclamp(float &value, const float &new_value) {
-      const float abs_value = fabs(new_value);
+      const float abs_value = float(fabs(new_value));
       if(value < abs_value)
         value = abs_value;
     }
@@ -258,7 +244,7 @@ namespace Zeni {
 
       const Vector3f valid_axes_distance = lhs.get_end_point_a() + min_max * lhs.get_direction() - closest_point;
       const float valid_axes_distance2 = valid_axes_distance * valid_axes_distance;
-      const float total_distance = sqrt(invalid_axes_distance2 + valid_axes_distance2);
+      const float total_distance = float(sqrt(invalid_axes_distance2 + valid_axes_distance2));
 
       return std::make_pair(total_distance, min_max);
     }
@@ -294,7 +280,7 @@ namespace Zeni {
     }
 
     float Plane::shortest_distance(const Point3f &rhs) const {
-      return fabs((point - rhs) * normal);
+      return float(fabs((point - rhs) * normal));
     }
 
     float Plane::shortest_distance(const Sphere &rhs) const {
@@ -541,7 +527,7 @@ namespace Zeni {
         {
           const float &ra = a[i];
           const float rb = b[0] * abs_R[i][0] + b[1] * abs_R[i][1] + b[2] * abs_R[i][2];
-          const float t = fabs(T[i]);
+          const float t = float(fabs(T[i]));
 
           if(t > ra + rb)
             return t - (ra + rb);
@@ -550,7 +536,7 @@ namespace Zeni {
         {
           const float ra = a[0] * abs_R[0][i] + a[1] * abs_R[1][i] + a[2] * abs_R[2][i];
           const float &rb = b[i];
-          const float t = fabs(T[0] * R[0][i] + T[1] * R[1][i] + T[2] * R[2][i]);
+          const float t = float(fabs(T[0] * R[0][i] + T[1] * R[1][i] + T[2] * R[2][i]));
 
           if(t > ra + rb)
             return t - (ra + rb);
@@ -570,7 +556,7 @@ namespace Zeni {
         for(int i = 0, m = 1, n = 2; i < 3; ++i) {
           const float ra = a[u] * abs_R[v][i] + a[v] * abs_R[u][i];
           const float rb = b[m] * abs_R[j][n] + b[n] * abs_R[j][m];
-          const float t = fabs(T[u] * R[v][i] - T[v] * R[u][i]);
+          const float t = float(fabs(T[u] * R[v][i] - T[v] * R[u][i]));
 
           if(t > ra + rb)
             return t - (ra + rb);
@@ -648,4 +634,4 @@ namespace Zeni {
 
 }
 
-#include <Zeni/Global_Undef.h>
+#include <Zeni/Undefine.h>
